% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oenb_attributes.R
\name{oenb_attributes}
\alias{oenb_attributes}
\title{Get Attributes}
\usage{
oenb_attributes(id, pos, lang = "EN")
}
\arguments{
\item{id}{character specifying the ID of the dataset of interest.
See \code{\link{oenb_toc}} to obtain the required ID.}

\item{pos}{character specifying the position ID of the indicator of interest.
See \code{\link{oenb_dataset}} to obtain the required ID.}

\item{lang}{Preferred language of the output. Possible values are "DE" for
German and "EN" for English (default).}
}
\value{
A data frame with one row per available value of an attribute and the
columns
\describe{
  \item{\code{attribute_code}}{the code of the attribute, such as
  \code{"dval1"}. It is a name of the \code{attr} argument of
  \code{\link{oenb_data}}.}
  \item{\code{attribute}}{the description of the attribute.}
  \item{\code{value_code}}{the code of the value, such as \code{"AT"}. It is
  a value of the \code{attr} argument of \code{\link{oenb_data}}.}
  \item{\code{value}}{the description of the value.}
}
Each value of an attribute is listed once.
A data frame without rows is returned if no attributes are found.
\code{NULL} is returned if the web service is not available.
See \code{\link{oenb}} for the workflow and the conditions that are signalled.
}
\description{
Get potential attributes of individual series from the OeNB's data web service.
}
\examples{
\donttest{
series_attr <- oenb_attributes(id = "11", pos = "VDBFKBSC217000")
series_attr
}

}
