% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oenb_toc.R
\name{oenb_toc}
\alias{oenb_toc}
\title{Table of Contents}
\usage{
oenb_toc(lang = "EN")
}
\arguments{
\item{lang}{Preferred language of the output. Possible values are "DE" for
German and "EN" for English (default).}
}
\value{
A data frame with one row per data set and the columns
\describe{
  \item{\code{dataset_id}}{the ID of the data set, which is the \code{id}
  argument of the other functions.}
  \item{\code{description}}{the title of the data set.}
}
\code{NULL} is returned if the web service is not available.
See \code{\link{oenb}} for the workflow and the conditions that are signalled.
}
\description{
Downloads the table of contents of the OeNB's statistical data web service.
}
\examples{
\donttest{
toc <- oenb_toc()
toc
}

}
