% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oenb_attributes.R
\name{oenb_attributes}
\alias{oenb_attributes}
\title{Get Attributes}
\usage{
oenb_attributes(id, pos, lang = "EN")
}
\arguments{
\item{id}{character specifying the ID of the dataset of interest.
See \code{\link{oenb_toc}} to obtain the required ID.}

\item{pos}{character specifying the position ID of the indicator of interest.
See \code{\link{oenb_dataset}} to obtain the required ID.}

\item{lang}{Preferred language of the output. Possible values are "DE" for
German and "EN" for English (default).}
}
\value{
A data frame containing potential attributes of a series.
}
\description{
Get potential attributes of individual series from the OeNB's data webservice.
}
\examples{
series_attr <- oenb_attributes(id = "11", pos = "VDBFKBSC217000")
series_attr

}
