% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oenb_data.R
\name{oenb_data}
\alias{oenb_data}
\title{Download OeNB Data}
\usage{
oenb_data(id, pos, freq = NULL, attr = NULL, starttime = NULL,
  endtime = NULL, lang = "EN")
}
\arguments{
\item{id}{character specifying the ID of the dataset of interest.
See \code{\link{oenb_toc}} to obtain the required ID.}

\item{pos}{character vector specifying the position IDs of the indicators of interest.
See \code{\link{oenb_dataset}} to obtain the required ID.}

\item{freq}{Frequency of the data. Where available, possible values are
\code{"D"}, \code{"M"}, \code{"Q"}, \code{"H"}, \code{"A"}
for daily, monthly, quarterly, semi-annual and annual data.
See \code{\link{oenb_frequency}} to obtain possible choices.}

\item{attr}{A named vector of further attributes.
See \code{\link{oenb_attributes}} to obtain possible choices.}

\item{starttime}{character specifying the start of the series. See 'Details'.}

\item{endtime}{character specifying the end of the series. See 'Details'.}

\item{lang}{Preferred language of the output. Possible values are "DE" for
German and "EN" for English (default).}
}
\value{
A data frame.
}
\description{
Download data sets from the OeNB's data webservice \url{https://www.oenb.at/en/Statistics.html}.
}
\examples{

series <- oenb_data(id = "11", pos = "VDBFKBSC217000", freq = "M", attr = c("dval1" = "AT"))
series

}
