% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oenb_dataset.R
\name{oenb_dataset}
\alias{oenb_dataset}
\title{Content of OeNB Datasets}
\usage{
oenb_dataset(id, lang = "EN")
}
\arguments{
\item{id}{character specifying the ID of the dataset of interest.
See \code{\link{oenb_toc}} to obtain the required ID.}

\item{lang}{Preferred language of the output. Possible values are "DE" for
German and "EN" for English (default).}
}
\value{
A data frame containing the IDs and names of available indicators within a dataset.
}
\description{
Downloads a description of the contents of a specific dataset from the OeNB's data webservice.
}
\examples{
content <- oenb_dataset(id = 11)
content

}
