% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oenb_metadata.R
\name{oenb_metadata}
\alias{oenb_metadata}
\title{Get Metadata}
\usage{
oenb_metadata(id, pos, lang = "EN")
}
\arguments{
\item{id}{character specifying the ID of the dataset of interest.
See \code{\link{oenb_toc}} to obtain the required ID.}

\item{pos}{character specifying the position ID of the indicator of interest.
See \code{\link{oenb_dataset}} to obtain the required ID.}

\item{lang}{Preferred language of the output. Possible values are "DE" for
German and "EN" for English (default).}
}
\value{
A data frame containing metadata on an indicator.
\code{NULL} is returned if the web service is not available.
}
\description{
Get metadata on individual series from the OeNB's data web service.
}
\examples{
\donttest{
meta <- oenb_metadata(id = "11", pos = "VDBFKBSC217000")
meta
}

}
