% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oenb_toc.R
\name{oenb_toc}
\alias{oenb_toc}
\title{Table of Contents}
\usage{
oenb_toc(lang = "EN")
}
\arguments{
\item{lang}{Preferred language of the output. Possible values are "DE" for
German and "EN" for English (default).}
}
\value{
A data frame containing the IDs and titles of available datasets.
}
\description{
Downloads the table of contents of the OeNB's statistical data web service.
}
\examples{
\dontrun{
toc <- oenb_toc()
toc
}

}
